requirements.txt
absl-py==2.3.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
dm-tree==0.1.9
docstring_parser==0.17.0
einops==0.8.1
etils==1.13.0
executing==2.2.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
h5py==3.14.0
idna==3.10
immutabledict==4.2.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
nest-asyncio==1.6.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.16.0
packaging==20.9
pandas==2.2.3
parso==0.8.4
pillow==10.3.0
platformdirs==4.3.6
promise==2.3
prompt_toolkit==3.0.50
protobuf==5.29.5
psutil==6.1.1
pure_eval==0.2.3
pyarrow==21.0.0
pygame==2.5.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
pyzmq==26.2.1
requests==2.32.4
rich==14.0.0
scikit-learn==1.7.1
scipy==1.15.2
seaborn==0.13.2
setuptools==80.9.0
simple-parsing==0.1.7
six==1.17.0
stack-data==0.6.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-datasets==4.9.9
tensorflow-hub==0.12.0
tensorflow-metadata==1.17.2
tensorflowjs==3.18.0
termcolor==3.1.0
threadpoolctl==3.6.0
tk==0.1.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.1
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zipp==3.23.0
